+++ frontend/src/components/layout/Header/hooks/index.ts
@@ -1,0 +1,5 @@
+import useBlockRefresh from './useBlockRefresh';
+import useExit from './useExit';
+import useRoutePath from './useRoutePath';
+
+export { useBlockRefresh, useExit, useRoutePath };
