+++ frontend/src/hooks/useDefaultMutationErrorHandler.ts
@@ -1,7 +1,7 @@
 import AlertModal from '@/components/AlertModal/AlertModal';
-import useModal from '@/hooks/useModal';
-import useToast from '@/hooks/useToast';
 import { CustomError, NetworkError } from '@/utils/error';
+
+import { useModal, useToast } from '@/hooks';
 
 const useDefaultMutationErrorHandler = () => {
   const { showToast } = useToast();
