+++ frontend/src/pages/ReadyPage/components/ReadyMembersContainer/ReadyMembersContainer.tsx
@@ -18,8 +18,8 @@ import SillyDdangkongMedium from '@/assets/images/sillyDdangkongMedium.webp';
 import A11yOnly from '@/components/common/a11yOnly/A11yOnly';
 import InviteModal from '@/components/InviteModal/InviteModal';
 import { QUERY_KEYS } from '@/constants/queryKeys';
-import { useGetRoomInfo } from '@/hooks/useGetRoomInfo';
-import useModal from '@/hooks/useModal';
+
+import { useGetRoomInfo, useModal } from '@/hooks';
 
 const ReadyMembersContainer = () => {
   const { members, master } = useGetRoomInfo();
