+++ frontend/src/pages/ReadyPage/components/RoomSetting/RoomSetting.tsx
@@ -11,7 +11,7 @@ import {
 
 import A11yOnly from '@/components/common/a11yOnly/A11yOnly';
 import RoomSettingModal from '@/components/RoomSettingModal/RoomSettingModal';
-import { useGetRoomInfo } from '@/hooks/useGetRoomInfo';
+import useGetRoomInfo from '@/hooks/useGetRoomInfo';
 import useIsMaster from '@/hooks/useIsMaster';
 import useModal from '@/hooks/useModal';
