+++ frontend/src/pages/RoundResultPage/RoundResultPage.tsx
@@ -1,5 +1,4 @@
-import NextRoundButton from './components/NextRoundButton/NextRoundButton';
-import RoundVoteContainer from './components/RoundVoteContainer/RoundVoteContainer';
+import { NextRoundButton, RoundVoteContainer } from './components';
 
 import Content from '@/components/layout/Content/Content';
