+++ frontend/src/pages/RoundResultPage/components/index.ts
@@ -1,0 +1,4 @@
+import NextRoundButton from './NextRoundButton/NextRoundButton';
+import RoundVoteContainer from './RoundVoteContainer/RoundVoteContainer';
+
+export { NextRoundButton, RoundVoteContainer };
